+++ src/utils.js
@@ -126,36 +126,6 @@ export function transformAdServerTargetingObj(targeting) {
 }
 
 /**
- * Read an adUnit object and return the sizes used in an [[728, 90]] format (even if they had [728, 90] defined)
- * Preference is given to the `adUnit.mediaTypes.banner.sizes` object over the `adUnit.sizes`
- * @param {object} adUnit one adUnit object from the normal list of adUnits
- * @returns {Array.<number[]>} array of arrays containing numeric sizes
- */
-export function getAdUnitSizes(adUnit) {
-  if (!adUnit) {
-    return;
-  }
-
-  let sizes = [];
-  if (adUnit.mediaTypes && adUnit.mediaTypes.banner && Array.isArray(adUnit.mediaTypes.banner.sizes)) {
-    let bannerSizes = adUnit.mediaTypes.banner.sizes;
-    if (Array.isArray(bannerSizes[0])) {
-      sizes = bannerSizes;
-    } else {
-      sizes.push(bannerSizes);
-    }
-  // TODO - remove this else block when we're ready to deprecate adUnit.sizes for bidders
-  } else if (Array.isArray(adUnit.sizes)) {
-    if (Array.isArray(adUnit.sizes[0])) {
-      sizes = adUnit.sizes;
-    } else {
-      sizes.push(adUnit.sizes);
-    }
-  }
-  return sizes;
-}
-
-/**
  * Parse a GPT-Style general size Array like `[[300, 250]]` or `"300x250,970x90"` into an array of sizes `["300x250"]` or '['300x250', '970x90']'
  * @param  {(Array.<number[]>|Array.<number>)} sizeObj Input array or double array [300,250] or [[300,250], [728,90]]
  * @return {Array.<string>}  Array of strings like `["300x250"]` or `["300x250", "728x90"]`
